+++ Shared/ForChildren/Stories/Detail/Data/StoryMetadata.swift
