+++ Shared/ForChildren/Stories/Detail/Player/AudioError.swift
@@ -1,0 +1,10 @@
+//
+//  AudioError.swift
+//  Movapp (iOS)
+//
+//  Created by Jakub Ruzicka on 21.11.2022.
+//
+
+struct AudioError: Error {
+    let message: String
+}
