+++ Shared/ForChildren/Stories/Detail/Presentation/StoryViewModel.swift
@@ -5,50 +5,8 @@
 //  Created by Jakub Ruzicka on 31.10.2022.
 //
 
-import AVFoundation
 import Combine
 import Foundation
-import UIKit
-
-enum PlayerState {
-    case none
-    case playing
-    case paused
-}
-
-struct PlayerContent {
-    let timer: PlayerTimer
-    let state: PlayerState
-    let languages: (selected: Languages, second: Languages)
-}
-
-struct PlayerTimer {
-    let currentTime: Double
-    let maxTime: Double
-}
-
-struct SentenceContent: Hashable {
-    let text: String
-    let isCurrent: Bool
-    let start: Double
-    let end: Double
-}
-
-struct StoryItemContent: Equatable {
-    let sentences: [SentenceContent]
-}
-
-struct StoryHeadline {
-    let image: String
-    let title: String
-    let subtitle: String
-}
-
-struct StoryContent {
-    let headline: StoryHeadline
-    let player: PlayerContent
-    let story: StoryItemContent
-}
 
 enum StoryState {
     case loading
@@ -209,135 +167,4 @@ class StoryViewModel: StoryViewModeling {
 
         currentTimeTimer?.invalidate()
     }
-}
-
-class StoriesAudioPlayers {
-
-    private let mainPlayer: StoryAudioPlayer
-    private let secondPlayer: StoryAudioPlayer
-
-    private var currentPlayer: StoryAudioPlayer
-
-    var duration: TimeInterval {
-        currentPlayer.duration
-    }
-
-    var currentTime: TimeInterval {
-        currentPlayer.currentTime
-    }
-
-    var isPlaying: Bool {
-        currentPlayer.isPlaying
-    }
-
-    init(slug: String, main: Languages, second: Languages) throws {
-        mainPlayer = try StoryAudioPlayer(suffix: "\(slug)/\(main.rawValue)")
-        secondPlayer = try StoryAudioPlayer(suffix: "\(slug)/\(second.rawValue)")
-
-        self.currentPlayer = mainPlayer
-    }
-
-    func switchLanguage() {
-        if currentPlayer == mainPlayer {
-            currentPlayer = secondPlayer
-        } else {
-            currentPlayer = mainPlayer
-        }
-    }
-
-    func play() {
-        currentPlayer.play()
-    }
-
-    func pause() {
-        currentPlayer.pause()
-    }
-
-    func stop() {
-        currentPlayer.stop()
-    }
-
-    func forward() {
-        currentPlayer.forward()
-    }
-
-    func backward() {
-        currentPlayer.backward()
-    }
-}
-
-class StoryAudioPlayer: Equatable {
-
-    private let seekingStep: TimeInterval = 10
-
-    private let avPlayer: AVAudioPlayer
-    private var observer: NSKeyValueObservation?
-    private var sessionCategory: AVAudioSession.Category?
-
-    var duration: TimeInterval {
-        avPlayer.duration
-    }
-
-    var currentTime: TimeInterval {
-        avPlayer.currentTime
-    }
-
-    var isPlaying: Bool {
-        avPlayer.isPlaying
-    }
-
-    // MARK: - Initializer
-
-    init(suffix: String) throws {
-        guard let data = NSDataAsset(name: "data/stories/\(suffix)") else {
-            throw AudioError(message: "File does not exists \(suffix)")
-        }
-
-        avPlayer = try AVAudioPlayer(data: data.data)
-    }
-
-    // MARK: - Functions
-
-    func play() {
-        sessionCategory = AVAudioSession.sharedInstance().category
-        try? AVAudioSession.sharedInstance().setCategory(AVAudioSession.Category.playback)
-
-        avPlayer.play()
-    }
-
-    func pause() {
-        if let sessionCategory = sessionCategory {
-            try? AVAudioSession.sharedInstance().setCategory(sessionCategory)
-        }
-
-        avPlayer.pause()
-    }
-
-    func stop() {
-        avPlayer.stop()
-    }
-
-    func forward() {
-        if avPlayer.currentTime + seekingStep > avPlayer.duration {
-            avPlayer.currentTime = avPlayer.duration
-        } else {
-            avPlayer.currentTime += seekingStep
-        }
-    }
-
-    func backward() {
-        if avPlayer.currentTime < seekingStep {
-            avPlayer.currentTime = 0
-        } else {
-            avPlayer.currentTime -= 10
-        }
-    }
-
-    static func == (lhs: StoryAudioPlayer, rhs: StoryAudioPlayer) -> Bool {
-        lhs.avPlayer == rhs.avPlayer
-    }
-}
-
-struct AudioError: Error {
-    let message: String
 }
